refactor: 졸업요건 타입과 라벨 공통 패키지로 이동
