--- packages/admin/src/components/graduation/lib/mappers.ts
@@ -1,33 +1,0 @@
-import type { CategoryType, ScopeType } from '@/hooks/server/graduation/useAdminGraduationView';
-
-export const CATEGORY_TYPE_LABELS: Record<CategoryType, string> = {
-  COMMON_REQUIRED: '공통교양',
-  BALANCE_REQUIRED: '균형교양',
-  ACADEMIC_BASIC: '학문기초',
-  GENERAL_ELECTIVE: '교양선택',
-  MAJOR_REQUIRED: '전공필수',
-  MAJOR_ELECTIVE: '전공선택',
-  MAJOR_BASIC: '전공기초',
-  TOTAL_COMPLETION: '총이수학점',
-  GENERAL: '교양',
-};
-
-export const SCOPE_TYPE_LABELS: Record<ScopeType, string> = {
-  PRIMARY: '주전공',
-  SECONDARY: '복수전공',
-  MINOR: '부전공',
-};
-
-export const BALANCE_AREA_LABELS: Record<string, string> = {
-  HISTORY_THOUGHT: '역사와 사상',
-  CULTURE_ARTS: '문화와 예술',
-  ECONOMY_SOCIETY: '경제와 사회',
-  NATURE_SCIENCE: '자연과 과학',
-};
-
-export const CLASSIC_DOMAIN_LABELS: Record<string, string> = {
-  WESTERN_HISTORY_THOUGHT: '서양의역사와사상',
-  EASTERN_HISTORY_THOUGHT: '동양의역사와사상',
-  EAST_WEST_LITERATURE: '동·서양의 문학',
-  SCIENCE_THOUGHT: '과학 사상',
-};
