+++ packages/client/src/features/graduation/ui/dashboard/CertificationSection.tsx
@@ -1,7 +1,6 @@
 import { useState } from 'react';
 import { Card, Flex, Grid, Badge, Button, Heading, SupportingText } from '@allcll/allcll-ui';
-import type { Certifications, ClassicDomain } from '@/entities/graduation/api/graduation';
-import { CLASSIC_DOMAIN_LABELS } from '../../lib/mappers';
+import { CLASSIC_DOMAIN_LABELS, type Certifications, type ClassicDomain } from '@allcll/common';
 import CertificationCriteriaModal from './CertificationCriteriaModal';
 
 export type CertificationType = 'english' | 'classic' | 'coding';
