+++ packages/common/src/lib/graduation/mappers.ts
@@ -1,6 +1,5 @@
-import type { BalanceRequiredArea, CategoryType, ClassicDomainType, ScopeType } from '@/entities/graduation/api/graduation';
+import type { BalanceRequiredArea, CategoryType, ClassicDomainType, ScopeType } from '../../types/graduation';
 
-/** 카테고리 타입 → 한글 라벨 매핑 */
 export const CATEGORY_TYPE_LABELS: Record<CategoryType, string> = {
   COMMON_REQUIRED: '공통교양',
   BALANCE_REQUIRED: '균형교양',
@@ -13,7 +12,6 @@ export const CATEGORY_TYPE_LABELS: Record<CategoryType, string> = {
   GENERAL: '교양',
 };
 
-/** 고전독서 도메인 → 한글 라벨 매핑 */
 export const CLASSIC_DOMAIN_LABELS: Record<ClassicDomainType, string> = {
   WESTERN_HISTORY_THOUGHT: '서양의역사와사상',
   EASTERN_HISTORY_THOUGHT: '동양의역사와사상',
@@ -21,14 +19,12 @@ export const CLASSIC_DOMAIN_LABELS: Record<ClassicDomainType, string> = {
   SCIENCE_THOUGHT: '과학 사상',
 };
 
-/** 스코프 타입 → 한글 라벨 매핑 */
 export const SCOPE_TYPE_LABELS: Record<ScopeType, string> = {
   PRIMARY: '주전공',
   SECONDARY: '복수전공',
   MINOR: '부전공',
 };
 
-/** 균형교양 영역 → 한글 라벨 매핑 */
 export const BALANCE_AREA_LABELS: Record<BalanceRequiredArea, string> = {
   HISTORY_THOUGHT: '역사와 사상',
   CULTURE_ARTS: '문화와 예술',
@@ -36,12 +32,10 @@ export const BALANCE_AREA_LABELS: Record<BalanceRequiredArea, string> = {
   NATURE_SCIENCE: '자연과 과학',
 };
 
-/** 이수 상태 라벨 */
 export function getStatusLabel(satisfied: boolean): string {
   return satisfied ? '이수 완료' : '미이수';
 }
 
-/** 이수 상태 뱃지 variant */
 export function getStatusBadgeVariant(satisfied: boolean): 'success' | 'danger' {
   return satisfied ? 'success' : 'danger';
 }
